pyproject.toml
[tool.poetry]
name = "wait-web"
version = "0.1.0"
description = "Open-source commerce website for local upcycle brand 'WAIT!'"
authors = ["roaddust2 <roaddust2@yahoo.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
psycopg2-binary = "2.9.6"
python-dotenv = "1.0.0"
gunicorn = "20.1.0"
pillow = "10.0.0"
django-modeltranslation = "0.18.9"
django-inline-svg = "0.1.1"
django-cleanup = "7.0.0"
django-allauth = "0.58.0"
django-bootstrap5 = "23.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
coverage = "^7.2.2"
django-debug-toolbar = "^4.1.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"